refactor: unify PWA and Desktop UI components via PlatformContext
Eliminates ~2,300 lines of duplicated component code between PWA and
Desktop packages by making PWA the canonical source for all shared UI.

Architecture:
- PlatformContext injects platform-specific store, capture functions,
  and layout slot components (TitleBar, SidebarConnectionSection,
  SourceIndicator, HeaderSyncIndicator, SettingsExtraSections)
- Shared types (BaseAppState, FilterOptions, ImportProgress) and
  focus-text utility moved to @freed/shared
- Desktop App.tsx now imports AppShell, FeedView, etc. from @freed/pwa

Deleted 11 desktop component files, 2 focus-text copies.
Created 5 desktop widgets, 2 PWA widgets, PlatformContext + types.

Made-with: Cursor
