--- packages/desktop/src/components/feed/FeedItem.tsx
@@ -1,1 +1,0 @@
-export { FeedItem } from "@freed/pwa/components/feed";
