--- packages/desktop/src/components/feed/FeedList.tsx
@@ -1,1 +1,0 @@
-export { FeedList } from "@freed/pwa/components/feed";
