--- packages/desktop/src/components/layout/Header.tsx
@@ -1,151 +1,0 @@
-import { useState, useMemo } from "react";
-import { AddFeedDialog } from "../AddFeedDialog";
-import { refreshAllFeeds } from "../../lib/capture";
-import { useAppStore } from "../../lib/store";
-
-interface HeaderProps {
-  onMenuClick: () => void;
-}
-
-export function Header({ onMenuClick }: HeaderProps) {
-  const [addFeedOpen, setAddFeedOpen] = useState(false);
-  const isSyncing = useAppStore((s) => s.isSyncing);
-  const feeds = useAppStore((s) => s.feeds);
-  const items = useAppStore((s) => s.items);
-  const markAllAsRead = useAppStore((s) => s.markAllAsRead);
-  const activeFilter = useAppStore((s) => s.activeFilter);
-  const feedCount = Object.keys(feeds).length;
-
-  const unreadCount = useMemo(
-    () =>
-      items.filter(
-        (item) =>
-          !item.userState.readAt &&
-          !item.userState.hidden &&
-          !item.userState.archived &&
-          (!activeFilter.platform || item.platform === activeFilter.platform) &&
-          (!activeFilter.savedOnly || item.userState.saved),
-      ).length,
-    [items, activeFilter],
-  );
-
-  const handleRefresh = async () => {
-    await refreshAllFeeds();
-  };
-
-  return (
-    <>
-      <header
-        className="h-12 flex items-center pr-4 border-b border-[rgba(255,255,255,0.08)] bg-[#0a0a0a]/80 backdrop-blur-xl"
-        data-tauri-drag-region
-        style={{ WebkitAppRegion: "drag" } as React.CSSProperties}
-      >
-        {/* Mobile menu button */}
-        <button
-          onClick={onMenuClick}
-          className="md:hidden p-2 ml-2 rounded-lg hover:bg-white/10 transition-colors"
-          aria-label="Open menu"
-          style={{ WebkitAppRegion: "no-drag" } as React.CSSProperties}
-        >
-          <svg
-            className="w-6 h-6"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M4 6h16M4 12h16M4 18h16"
-            />
-          </svg>
-        </button>
-
-        {/* Spacer */}
-        <div className="flex-1" />
-
-        {/* Actions */}
-        <div
-          className="flex items-center gap-2"
-          style={{ WebkitAppRegion: "no-drag" } as React.CSSProperties}
-        >
-          {/* Mark all read — only shown when there are unread items */}
-          {unreadCount > 0 && (
-            <button
-              onClick={() => markAllAsRead(activeFilter.platform)}
-              title={`Mark all ${unreadCount} items as read`}
-              className="hidden sm:flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-sm text-[#71717a] hover:bg-white/5 hover:text-white transition-colors"
-            >
-              <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
-              </svg>
-              <span>{unreadCount} unread</span>
-            </button>
-          )}
-
-          {/* Refresh button */}
-          {feedCount > 0 && (
-            <button
-              onClick={handleRefresh}
-              disabled={isSyncing}
-              className="p-2 rounded-lg hover:bg-white/10 transition-colors disabled:opacity-50"
-              aria-label="Refresh feeds"
-            >
-              <svg
-                className={`w-5 h-5 ${isSyncing ? "animate-spin" : ""}`}
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                />
-              </svg>
-            </button>
-          )}
-
-          {/* Add feed button */}
-          <button
-            onClick={() => setAddFeedOpen(true)}
-            className="flex items-center gap-2 px-3 py-1.5 rounded-lg bg-[#8b5cf6]/20 text-[#8b5cf6] hover:bg-[#8b5cf6]/30 transition-colors"
-          >
-            <svg
-              className="w-4 h-4"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M12 4v16m8-8H4"
-              />
-            </svg>
-            <span className="text-sm font-medium hidden sm:inline">
-              Add Feed
-            </span>
-          </button>
-
-          {/* Sync status */}
-          <div className="flex items-center gap-2 text-sm text-[#71717a] ml-2">
-            <span
-              className={`sync-dot ${
-                isSyncing ? "syncing" : "connected"
-              }`}
-            />
-            <span className="hidden sm:inline">
-              {isSyncing ? "Syncing..." : feedCount > 0 ? "Synced" : "Ready"}
-            </span>
-          </div>
-        </div>
-      </header>
-
-      <AddFeedDialog open={addFeedOpen} onClose={() => setAddFeedOpen(false)} />
-    </>
-  );
-}
