+++ packages/pwa/package.json
@@ -4,7 +4,11 @@
   "version": "0.1.0",
   "type": "module",
   "exports": {
-    "./components/feed": "./src/components/feed/index.ts"
+    "./components/feed": "./src/components/feed/index.ts",
+    "./components/layout": "./src/components/layout/index.ts",
+    "./components/AddFeedDialog": "./src/components/AddFeedDialog.tsx",
+    "./components/SettingsPanel": "./src/components/SettingsPanel.tsx",
+    "./context": "./src/context/index.ts"
   },
   "scripts": {
     "dev": "vite",
