+++ packages/pwa/src/components/feed/index.ts
@@ -1,2 +1,4 @@
 export { FeedList } from "./FeedList";
 export { FeedItem } from "./FeedItem";
+export { FeedView } from "./FeedView";
+export { ReaderView } from "./ReaderView";
