+++ packages/pwa/src/components/layout/AppShell.tsx
@@ -1,4 +1,5 @@
 import { useState, type ReactNode } from "react";
+import { usePlatform } from "../../context/PlatformContext";
 import { Sidebar } from "./Sidebar";
 import { Header } from "./Header";
 
@@ -7,10 +8,14 @@ interface AppShellProps {
 }
 
 export function AppShell({ children }: AppShellProps) {
+  const { TitleBar } = usePlatform();
   const [sidebarOpen, setSidebarOpen] = useState(false);
 
   return (
     <div className="flex-1 min-h-0 flex flex-col bg-[#121212]">
+      {/* Optional platform title bar (e.g. Tauri chrome with drag region) */}
+      {TitleBar && <TitleBar />}
+
       {/* Header - fixed height */}
       <Header onMenuClick={() => setSidebarOpen(true)} />
 
@@ -20,7 +25,7 @@ export function AppShell({ children }: AppShellProps) {
         <Sidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} />
 
         {/* Main content */}
-        <main className="flex-1 h-full overflow-hidden">{children}</main>
+        <main className="flex-1 min-h-0 overflow-hidden">{children}</main>
       </div>
     </div>
   );
