+++ packages/pwa/src/components/layout/index.ts
@@ -1,0 +1,1 @@
+export { AppShell } from "./AppShell";
