+++ packages/shared/src/focus-text.ts
