+++ packages/shared/src/index.ts
@@ -13,6 +13,12 @@ export * from "./ranking.js";
 // Re-export OPML utilities (browser-safe, no Automerge)
 export * from "./opml.js";
 
+// Re-export focus text utilities (browser-safe, no deps)
+export * from "./focus-text.js";
+
+// Re-export store types (browser-safe, no deps)
+export * from "./store-types.js";
+
 // Note: schema.js is NOT re-exported here because it imports Automerge
 // which uses WebAssembly and requires special bundler configuration.
 // For schema operations, import directly from '@freed/shared/schema' in Node.js/Tauri.
